Moved the home page to its own package
